refactor: extract MenuActions into its own file
Move the MenuActions class (~2,450 lines) from ContextMenuBuilder.swift
into a dedicated MenuActions.swift file. This is a pure file split with
zero logic changes — ContextMenuBuilder.swift drops from 4,949 to 2,497
lines, making both files easier to navigate and maintain.

The helper structs (SonosRoomToggle, SonosGroupingInfo, SonosRoomAction)
remain in ContextMenuBuilder.swift since they are used by both files and
are visible module-wide.

Closes #164
